AuthorizationService/requirements.txt
fastapi==0.115.0
requests==2.32.3
uvicorn==0.31.0
django-environ==0.11.2
setuptools==75.5.0
pydantic-settings==2.6.1
pytest==8.3.3
Faker==33.0.0